GPLv3 entire project, and adding specific headers where it makes sense.
The entire package is licensed from here onwards, under GPLv3. Boilerplate per file is a reminder only, and does not make sense adding it everywhere liberally, both for space reasons, and that the opensource community are far from being idiots. Also, where the code was generated by lumen, and non-specific to the project, no boilerplate was added. Thank you gents and ladies of the code world.